+++ src/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Route, Switch } from "react-router-dom";
 import "./App.css";
 // * Component Imports
-import RoutesTree from "./Components/RoutesTree";
+import RoutesTree from "./RoutesTree";
 import Header from "./Components/HeaderComponents/Header";
 import NavBar from "./Components/NavBarComponents/NavBar";
 import Footer from "./Components/SharedComponents/Footer";
