+++ src/RoutesTree.jsx
@@ -2,24 +2,24 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 
 // * Component Imports
-import Home from './Home/Home';
-import Signup from './Signup/Signup';
-import Login from './Login/Login';
-import ResetPassword from './ResetPassword/ResetPassword';
-import ResultsPage from './Results/ResultsPage';
-import AboutPage from './About/About';
-import ContactPage from './Contact/Contact';
+import Home from './Pages/Home';
+import Signup from './Pages/Signup';
+import Login from './Pages/Login';
+import ResetPassword from './Pages/ResetPassword';
+import ResultsPage from './Pages/ResultsPage';
+import AboutPage from './Pages/About';
+import ContactPage from './Pages/Contact';
 // import DanceStyles from "./DanceStyles/DanceStyles";
-import BreakingInfo from './DanceStyles/BreakingInfo';
-import HipHopInfo from './DanceStyles/HipHopInfo';
-import HouseInfo from './DanceStyles/HouseInfo';
-import LockingInfo from './DanceStyles/LockingInfo';
-import PoppingInfo from './DanceStyles/PoppingInfo';
-import VogueingInfo from './DanceStyles/VogueingInfo';
-import WaackingInfo from './DanceStyles/WaackingInfo';
-import FindEvents from './EventComponents/FindEvents';
-import RecentEvents from './EventComponents/RecentEvents';
-import SubmitEventInfo from './EventComponents/SubmitEventInfo';
+import BreakingInfo from './Components/DanceStyles/BreakingInfo';
+import HipHopInfo from './Components/DanceStyles/HipHopInfo';
+import HouseInfo from './Components/DanceStyles/HouseInfo';
+import LockingInfo from './Components/DanceStyles/LockingInfo';
+import PoppingInfo from './Components/DanceStyles/PoppingInfo';
+import VogueingInfo from './Components/DanceStyles/VogueingInfo';
+import WaackingInfo from './Components/DanceStyles/WaackingInfo';
+import FindEvents from './Components/EventComponents/FindEvents';
+import RecentEvents from './Components/EventComponents/RecentEvents';
+import SubmitEventInfo from './Components/EventComponents/SubmitEventInfo';
 
 const RoutesTree = () => {
     return (
